Fixed issue where container was too narrow
I felt like the container looked too narrow on my 1920x1200 display,
and found it was at 730px which seemed wrong. I found this code that
was setting the container's max width to 730px. Once I removed this,
the container was able to size to 970px like it's supposed to.